airflow/requirements.txt
podaac-data-subscriber
earthaccess
h5py
geoalchemy2
gcsfs
google-cloud-storage